+++ src/App.js
@@ -2,16 +2,15 @@ import { Container } from "react-bootstrap";
 import "./App.scss";
 import { ContextProvider } from "./model/ContextModel";
 import Header from "./templates/Header";
-import LRlevel2Grid from "./templates/LRlevel2Grid";
+import HrManagement from "./templates/HrManagement";
 
 function App() {
   return (
     <div>
       <ContextProvider>
         <Header />
-        <Container>
-          <LRlevel2Grid />
-        </Container>
+
+        <HrManagement />
       </ContextProvider>
     </div>
   );
