+++ src/model/ContextModel.js
@@ -13,7 +13,6 @@ const ContextProvider = ({ children }) => {
   const url = "http://localhost:8888"; // REST API 서버 주소
 
   const [selectedRows, setSelectedRows] = useState([]);
-  const [reloadSubTableData, setReloadSubTableData] = useState(false);
 
   const deleteSelectedRows = useCallback(() => {
     // 각 row에 대한 delete 요청을 생성
@@ -31,7 +30,6 @@ const ContextProvider = ({ children }) => {
       .then((responses) => {
         console.log("선택된 모든 행의 삭제 완료");
         setSelectedRows([]); // 선택행 배열 비우기
-        setReloadSubTableData(!reloadSubTableData);
       })
       .catch((error) => {
         console.error("하나 이상의 요청에서 에러 발생: ", error);
@@ -40,7 +38,7 @@ const ContextProvider = ({ children }) => {
   }, [selectedRows]);
 
   const value = {
-    contextState: { selectedRows, reloadSubTableData },
+    contextState: { selectedRows },
     contextActions: {
       setSelectedRows,
       deleteSelectedRows,
