+++ src/templates/HrManagementHeader.js
@@ -1,0 +1,64 @@
+import { faTrashCan } from "@fortawesome/free-regular-svg-icons";
+import {
+  faArrowUpRightFromSquare,
+  faBorderAll,
+  faCalculator,
+  faPrint,
+} from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import React, { useContext, useState } from "react";
+import { Button } from "react-bootstrap";
+import "../styles/header.css";
+import empAdd from "../styles/img/empAddLogo.png";
+import ContextModel from "../model/ContextModel";
+
+const HrManagementHeader = () => {
+  const [showSidebar, setShowSidebar] = useState(false);
+
+  const { contextActions } = useContext(ContextModel);
+
+  const toggleSidebar = () => {
+    setShowSidebar(!showSidebar);
+  };
+
+  return (
+    <div id="secondTopHeader">
+      <div id="secondTopHeaderContents">
+        <Button
+          id="toggleSidebarBtn"
+          onClick={toggleSidebar}
+          variant="outline-secondary"
+        >
+          <i className={`fa fa-bars colorWhite`} />
+        </Button>
+        {/* 로고 */}
+        <img id="logo" src={empAdd} alt="" />
+        <button className="backgroundBorderNone">
+          <FontAwesomeIcon
+            icon={faArrowUpRightFromSquare}
+            className="colorWhite backgroundBorderNone"
+          />
+        </button>
+      </div>
+      <div id="secondTopHeaderMenuList">
+        <button className="backgroundBorderNone">
+          <FontAwesomeIcon icon={faPrint} className="colorWhite" />
+        </button>
+        <button
+          className="backgroundBorderNone"
+          onClick={(e) => contextActions.deleteSelectedRows()}
+        >
+          <FontAwesomeIcon icon={faTrashCan} className="colorWhite" />
+        </button>
+        <button className="backgroundBorderNone">
+          <FontAwesomeIcon icon={faCalculator} className="colorWhite" />
+        </button>
+        <button className="backgroundBorderNone">
+          <FontAwesomeIcon icon={faBorderAll} className="colorWhite" />
+        </button>
+      </div>
+    </div>
+  );
+};
+
+export default HrManagementHeader;
